// app/services/api/api.types.ts
import { GeneralApiProblem } from "./api-problem";

export interface User {
  id: number;
  name: string;
}

export type GetUsersResult = { kind: "ok"; users: User[] } | GeneralApiProblem;
export type GetUserResult = { kind: "ok"; user: User } | GeneralApiProblem;

export interface ApiResult<T> {
  kind: string;
  data?: T;
  error?: string;
}